requirements.txt
pip==23.1.2
requests==2.31.0
setuptools==68.0.0
python-slugify==8.0.1
mutagen~=1.46.0
pytube~=15.0.0
moviepy~=1.0.3
beautifulsoup4~=4.12.2
tqdm~=4.65.0
# selenium~=3.141.0 # webdriver compatible con Opera
selenium~=4.14.0